PVQ: Fix crash in average cdf when ec_smallmul is off
Some of symbols in PVQ decides its size of alphabet during runtime,
i.e. when it is first encoded, so the cdf is initialized when the symbol
is first encoded. However, some of those symbols can be never encoded
then the cdf is not initialized. This leads crash when averraging the
cdf from each tile and ec_smallmul is turned off.

Change-Id: I045e36ed5aec31dc65b8e6c132fcc2c6b7afdd19