Update patch set 1
Patch Set 1: Code-Review-2

This won't work because the signaling is conditional on sb_all_skip(), which can only be determined once all partitions have been received on the decoder.

The solution is to signal CDEF right after it's determined if the largest partition is skip or not, and change the sb_all_skip() condition to that. It will incur a small but acceptable coding loss.

Patch-set: 1
Reviewer: Gerrit User 5050 <5050@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=-2