Update patch set 1
Patch Set 1:

Jingning: Thanks for the comments! I believe I've tracked down the source of the test failures - my changes regarding what to do when we hit the maximum transform splitting depth were inconsistent between the encoder and decoder. That explains how mbmi->tx_size ended up with different values.

As you say, it might be worth trying out various options for the intra tx size contexts, but that would probably be best done as a separate patch since it doesn't seem to have been the immediate cause of the test failures.

Patch-set: 1