Make loop-restoration use 64x64 processing units
Changes loop-restoration to use processing unit size that is
64x64 for luma; for chroma the processing unit is coupled to
the same 64x64 support region as luma.
Thus for chroma the processing unit size is 32x32 for 4:2:0,
32x64 for 4:2:2 and 64x64 for 4:4:4, etc.

While the Wiener filter output should not change with this patch,
the sgr filter will change since the boundary pixel handling in
sgr is internal within the filter.

Change-Id: I65a9e2df88927a19445420ce400acb1fcf7afa93