Reduce line buffer size for Wiener filter
This patch forces the vertical filtering for the top and bottom rows
of a processing unit for the Wiener filter to be 5-tap.  The 5-taps
are derived from the primary 7-tap filter by forcing the taps at the
end to be zero, and absorbing their weights into the other taps to
maintain normalization.

This will effectively reduce the line buffer size for luma Wiener
filter to 4 (from 6).

Change-Id: I5e21b58369777eabf553a8987387d112f98a5598