Eliminate line buffer size for Wiener filter.
This patch forces the vertical filtering for the top and bottom
rows of a processing unit for the Wiener filter to not cross
boundaries of processing units.

This will effectively eliminate the line buffer requirement for
the Wiener filter.

The macro RESTORATION_BORDER_VERT can be used to control the
line buffer. Currently set at 0, the line buffer is completely
eliminated; but the macro can be changed to 1, 2 or 3 to test
other options.

Change-Id: Ie58a98c784a0db547627b9cfcf55f018c30e8e79