Eliminate line buffer size for Wiener filter.
This patch forces the vertical filtering for the top and bottom
rows of a processing unit for the Wiener filter to not cross
boundaries of processing units.

This will effectively eliminate the line buffer requirement for
the Wiener filter.

Also, for the sgr filter we add the option of using overlapping
windows horizontally to improve coding efficiency.

Change-Id: I7a595ebc195f661b73926975c7deee1f8c6daad3