Reduce/Eliminate line buffer for loop-restoration.
This patch forces the vertical filtering for the top and bottom
rows of a processing unit for the Wiener filter to not use border
more than what is set in the RESTORATION_BORDER_VERT macro.
This macro is currently set at 0 to eliminate line buffer completely,
but it could be increased to 1 or 2 to use limited line buffers
if the coding efficiency is affected too much with a 0 line-buffer.

Also, for the sgr filter we added the option of using overlapping
windows horizonttally and vertically to improve coding efficiency.
The vertical border used is set by the same RESTORATION_BORDER_VERT
macro, while the horizontal border can be set by the
RESTORATION_BORDER_HORZ macro set at 3. Currently we do not
recommend changing RESTORATION_BORDER_HORZ below 3, but the
RESTORATION_BORDER_VERT macro could be set as 0 (no line buffer),
1 (2-line buffer), 2 (4-line buffer), or 3 (6-line buffer) as
needed.

Change-Id: Ie58a98c784a0db547627b9cfcf55f018c30e8e79