Constrain motion vector projection range
Constrain the maximum motion vector projection range to be within
+/-32 pixels in the vertical direction and +/-64 pixels in the
horizontal direction.

Such constraints allow a fixed amount of reference motion vector
load to SRAM for each 64x64 block size, independent of the frame
size. The wider range in the horizontal direction can be stored in
the SRAM and reused by next 64x64 block. The compression performance
loss is 0.03% for lowres and 0.04% for midres.

Change-Id: I7f1c136363b136b1f2fa9f7c962a791c8e91a976