Round up subsampled frame size in av1_loop_restoration_corners_in_sb
The previous code converted a frame_w (say) of 1 to zero for a plane
where subsampling was enabled, causing a division by zero in
av1_get_rest_ntiles. This doesn't match the spec, which says
subsampling rounds up.

The patch adds the rounding, and also adds an assertion to
av1_get_rest_ntiles to help diagnose any other broken callsites.

Change-Id: Ia6c249fa935c3a16d122ba6e7b450fe99f412fde