Update patch set 4
Patch Set 4: Code-Review+2

> Patch Set 4:
> 
> > Patch Set 3:
> > 
> > I think allowing tile_rows and tile_cols to not be a power of 2 is reasonable. The way of deciding tile size in AV1 is different from that in VP9, which causes the problem you mentioned. There is another patch that has the similar idea implemented. Could you take a look?
> > 
> > https://aomedia-review.googlesource.com/c/aom/+/6468
> 
> That is a much more ambitious change, I think. This patch was just trying to give a (non-broken) meaning to something that can already be signalled.

Thanks. This should be fine for now. Later, we should allow the users to set tile_rows and tile_cols that are not a power of 2, and transmit the actual # of tiles in bitstream.

Patch-set: 4
Reviewer: Gerrit User 5185 <5185@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=+2