Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> (1 comment)

MSVC and Xcode do $CMAKE_C_FLAGS $CMAKE_C_FLAGS_<CONFIG>. Well, the versions I've tested with do that... I can try changing msvc's and cmake's.

AFAICT all generators behave this way. We have to append to the <CONFIG> variants to ensure the flags get obeyed over defaults in the <CONFIG> versions of the flags. In practice the <CONFIG> versions have values while the base is empty until/unless touched by our scripts.

This is why enabling asserts in release configs can be painful-- you have to set the <CONFIG> version of the flags and hope that the generator appended what was passed in the arg if you needed to override the NDEBUG passed in the default release configs.

The AOM_EXTRA_* stuff deals with this in an ugly'ish way, but it seems to work.

Patch-set: 6
Reviewer: Gerrit User 5215 <5215@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>