Update patch set 2
Patch Set 2:

It's good that you caught this.  I think this became possible when I decided to simplify the damping introducing shared luma/chroma signalling for singlepass (since separate signalling didn't seem to help compression).

Fortunately, as in the SIMD code, this doesn't doesn't add an extra operation per pixel, since the strength is the same for the entire block, so the patch looks perfectly OK, but to be sure that we don't miss better ways around this, let's think about this for a moment.

Other options could be:

1) To signal the damping separately for luma and chroma as it used to be at the cost of two bits per frame (and the encoder can simply signal max(luma_damping - 1, 3) for chroma (I think a proper RDO will give little gain)

2) We could have chroma_damping = luma_damping.  This might give a slight compression loss (it used to give that, but possibly within the noise range)

3) Mandate that chroma_damping is max(luma_damping - 1, 3), making 3 the lowest even for strengths < 8.

Personally, I think the patch is good as it is and the design doc can be updated accordingly.  But I'd like your opinions.

Patch-set: 2
Reviewer: Gerrit User 5111 <5111@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>