Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Thinking of 8-bit content again for simplicity, my understanding is
>  > that the primary strength can be signalled up to 15, but there is a
>  > special case where the primary strength is set to 19. Note that
>  > get_msb(19) = 4, so any damping < 4 is potentially an issue; in
>  > particular, that means you'd have to either adjust the range of
>  > luma dampings as well, or remove the strength=19 option.
>  
> It may seem so, but this is not the case.  The luma strength gets modified by adjust_strength() which will never return more than 15 for a strength of 19.  19 was carefully selected for that reason.
> 
> I don't really like this special case, but it makes no sense to have two ways to signal zero strengths either.

Okay - that makes sense. So the luma plane is fine, but the chroma plane also has that special case and doesn't call adjust_strength(). I've just double-checked and it is indeed possible to get to cdef_filter_block() with pri_strength = 19 for a chroma plane.

So we'd have to either bump the minimum chroma damping to 4, or to only apply the pri_strength = 19 special case to the luma plane.

Patch-set: 2
Reviewer: Gerrit User 5975 <5975@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>