Define coding parameters for GF group of 16
Propose the coding structure for GF Group of 16 to use the following
multi-layer design:
  1st Layer: Frame 0 and Frame 16 (ALTREF)
  2nd Layer: Frame 8 (ALTREF2)
  3rd Layer: Frame 4 and 12 (ALTREF2)
  4th Layer: Frame 2, 6, 10, and 14 (BWDREF)
  5th Layer: Frame 1, 3, 5, 7, 9, 11, 13, and 15

This CL implements the reference frame update for a GF group of 16
using a look-up table, which is different from the current
implementation. The motivation behind is to try out various GF group
coding structures simply through manipulating the look-up table,
instead of making changes on swapping the reference indexes and
associated flags throughout the codebase.

Change-Id: I823edae1c26263e0e0642a580685a25e81caa0a9