Initialize cm->prev_frame before decoding gm params
This fixes global motion read/write mismatch.
Minimal test case would be 3 encoded frames with:
    1) frame0: KEY frame
    2) frame1: INTER frame with refresh_frame_flags not 0x00 and not 0xFF, say 0x01 (refresh first slot)
    3) frame2: INTER frame with all reference indices equal to 3 (i.e. no reference to frame1)

When decoding frame2 decoder will use cm->prev_frame=frame1 for gm params decoding
then switch to cm->prev_frame=frame0 for the rest of the frame.
This makes frame2 dependent on frame1 which is not used for reference.
And also it is different from what encoder does.

BUG=aomedia:750

Change-Id: I42ba19c03d909286af12ba3149910693a12249e8