Improve dist-4x4 to compute 8x8 yuv dist for sub8x8
To apply dist-8x8 for sub8x8 partitions, once mode decision for
sub8x8 partitions are finished then dist-8x8 is computer on 8x8 window.
Since dist-8x8 is only for luma, chroma distortion should be identified.

Previously, it has been hard to be free from potential bugs doing this,
due to the complexity of inter mode search code.

The new and less-error-prone method is to compute uv distortion (in MSE) after
the mode decisions for all of sub8x8 blocks in a 8x8 window are finished,
when the dist-8x8 distortion for luma 8x8 pixels are computed with
new distortion metric.

Change-Id: Ieaccb7915df5faeb5e89a7e70b2b7cbac65231af