Update patch set 29
Patch Set 29:

(4 comments)

For single-tile encoding, I observe quite a big loss - around 0.8% for vidyo4 in low delay config, QP 42. Can we verify the source of this loss and whether it can be compensated for, e.g. in the IVF mapping?

There is a gain for multiple tiles and TGs but this comes from removing the frame header copy in the TGs, and it would be good to have sub-frame resilience supported before turning on by default.

Patch-set: 29
Reviewer: Gerrit User 5065 <5065@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>