Update patch set 29
Patch Set 29:

(4 comments)

> Patch Set 29:
> 
> (4 comments)
> 
> For single-tile encoding, I observe quite a big loss - around 0.8% for vidyo4 in low delay config, QP 42. Can we verify the source of this loss and whether it can be compensated for, e.g. in the IVF mapping?
> 
> There is a gain for multiple tiles and TGs but this comes from removing the frame header copy in the TGs, and it would be good to have sub-frame resilience supported before turning on by default.

The loss with single-tile encoding is due to the fact that I am setting aside 4 bytes at the beginning of EVERY obu (that is 12 bytes/frame for INTER frames) for storing the size of the obu.  Technically, the TD and sequence/frame header obus don't need this, but I made it this way because I think it would make more sense going forward, such as if/when we need to add start codes.

I think I'll put in an option to NOT write the 4-byte obu for those that don't need to

Patch-set: 29
Reviewer: Gerrit User 7963 <7963@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>