Update patch set 29
Patch Set 29:

> Patch Set 29:
> 
> (4 comments)
> 
> > Patch Set 29:
> > 
> > (4 comments)
> > 
> > For single-tile encoding, I observe quite a big loss - around 0.8% for vidyo4 in low delay config, QP 42. Can we verify the source of this loss and whether it can be compensated for, e.g. in the IVF mapping?
> > 
> > There is a gain for multiple tiles and TGs but this comes from removing the frame header copy in the TGs, and it would be good to have sub-frame resilience supported before turning on by default.
> 
> The loss with single-tile encoding is due to the fact that I am setting aside 4 bytes at the beginning of EVERY obu (that is 12 bytes/frame for INTER frames) for storing the size of the obu.  Technically, the TD and sequence/frame header obus don't need this, but I made it this way because I think it would make more sense going forward, such as if/when we need to add start codes.
> 
> I think I'll put in an option to NOT write the 4-byte obu for those that don't need to

I did a little more careful analysis with your coding scenario.  Coding vidyo4 with your conditions for 600 frames resulted in 810299 bytes before, and 819255 bytes with new code.  That is 8956 extra bytes, which is roughly 1% more (decoded yuvs are identical).  Of the 8956 extra, 12 (4 bytes for TD size, 4 bytes for Frame header size, and 4 bytes for TG size) are not part of the bitstream for the 599 INTER frames, and 16 bytes are not part of the bitstream for the one KEY frame.  Thus, 12*599+16=7204 bytes are not really part of the bitstream.  Thus, only 1752 bytes (8956-7204) are the effective increase resulting from HLS.  These are attributed to obu headers themselves (1 byte per OBU), new sequence header, etc.  So, the increase in bitrate is only 1752/810299 = 0.2%, which is to be expected.  Once a proper ivf/webm muxing is put in, the extra 7284 bytes should go away.

Patch-set: 29
Reviewer: Gerrit User 7963 <7963@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>