[dist-8x8] Enable early-exit, if MSE is used
If dist-8x8 is enabled, it has disabled early-exit during the RD
of sub8x8 tx block in 8x8 or larger partition for plane 0.

However, for the sanity check mode of dist-8x8 which still use MSE,
this can cause different bdrate from the reference (i.e
dist-8x8 off). Hence, if tune-metirc is MSE then
regular early-exit will perform for all cases.

Note that the rd cost used for early-exit condition during tx mode
decision does not include rate for mode itself but only coefficints.
So, when the total distortion for a partition is very small, the rd cost
of mode can exceed that of distortion.

Change-Id: Ie25c64064b78d2d1582dbb0c286212fbc52dbdb9