Change CDEF block skip condition and move signalling
Previously CDEF was implicitly disabled for a superblock if all its
sub-blocks were skip.  Now CDEF is implicitly disabled only if the
superblock is skip and not partitioned.  This means that it's no
longer necessary to decode the whole superblock before knowing how to
filter it, and CDEF is instead signalled after the skip bit of the top
left block whatever its size.  If the superblock size is 128x128, four
CDEF presets are signaled, one for each 64x64 block in regular raster
order (unless 128x128 block is skip and not partitioned).

After the introduction of the skip condition flag, this change also
makes it possible to filter a partitioned block even though all of it
is skipped.  Such superblocks are somewhat rare, and the quality
impact is within the noise range.

Change-Id: I3adfa3d2d01e288b7db876198aa9985ee9f53917