Change CDEF block skip condition and move signalling
Previously CDEF was implicitly disabled for a filter block if all its
sub-blocks were skip, and no bits for that block was signalled.  That
required the CDEF signal to be transmitted at the end of the block,
and it was not possible to begin the filtering of the block before
that.

This patch moves the signalling to just after the first zero skip bit
within a 64x64, 128x64, 64x128 or 128x128 block.  If sub-blocks are
skip, then no CDEF bits will be signalled, as before.  Also, the skip
condition flag has been removed so it's always known at the skip flag
whether a coding block is to be filtered or not.

Change-Id: I3adfa3d2d01e288b7db876198aa9985ee9f53917