Add an SSE4.1 implementation of av1_convolve_2d_scale
There are still some speed-ups possible (for small blocks, there's a
fallback to straight C code), but for large blocks this is almost 8x
the speed of the C version. The code needs SSE 4.1 for the PMULLD
instruction that we use to do SIMD 32-bit multiplies.

This patch also adds a test av1_convolve_scale_test to make sure the
optimised code matches the C version.

Change-Id: I6732bc6b2378ffaadae5aa6441100cf660f7ee11