WIP: Pack InterpFilters into a single integer
Rather than keeping an array of four 4-bit numbers (if
CONFIG_DUAL_FILTER), this patch changes the code to pack them into a
uint32_t. This is worth doing for two reasons. First, there are reams
of code that special-case with #if CONFIG_DUAL_FILTER that can be
folded into a single case after this change. Second, the resulting
code might actually be quicker or at least smaller: we very rarely
insert a value into the array (count the calls to
av1_update_interp_filters), but we often look up values from it.

This patch should be considered a work-in-progress: I've compiled with
and without --disable-dual-filter, but I assume there are many
combinations of flags which will fall over...

Change-Id: I6b24a61bac3e4b220d8d46d0b27cfe865dcfba81