Update patch set 32
Patch Set 32:

(1 comment)

You mean there is a bug with this patch for error resilient? Or a bug in master? And this is fixed in a patch on top of this one?

I think your fix is ok so long as this is WIP.

Patch-set: 32
Reviewer: Gerrit User 5065 <5065@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>