Write to correct buffer in av1_make_masked_inter_predictor
With CONVOLVE_OPT_NO_ROUND and a masked compound type,
build_inter_predictors generates two predictions as follows. It first
calls av1_make_inter_predictor for the first ref, telling it to write
to a "tmp_dst" on the stack frame of build_inter_predictors. It then
calls av1_make_masked_inter_predictor for the second ref. This
generates a prediction by calling av1_make_inter_predictor and should
write the results to a second (32-bit) temporary buffer. These two
buffers are finally blended with a mask using build_masked_compound,
writing to the frame in "dst".

Before this patch, the second call to av1_make_inter_predictor
overwrote the first tmp_dst (it writes to conv_params->dst). Then the
mask was computed by taking a difference between that and the garbage
in tmp_dst2[*].

This patch moves the declaration of tmp_dst2 and the code that
overrides conv_params->dst to come before the call to
av1_make_inter_predictor. As a result, the compound mode should
actually be able to blend between both references.

[*] It has only been garbage since my commit 07089c6 which mistakenly
    removed a memset. Before that change, the mask was taking a
    difference between the second reference frame and zero, but it was
    still wrong.

Change-Id: I342bdd56cf3eded11f8c975253c315740bd0e46e