Update patch set 7
Patch Set 7:

Hi James,

Where are we with this patch? As it stands, the patch contains an optimised SSE version of an existing C function. It also adds a new test which checks that the SSE code does the same thing as the C function.

You've asked for a C only test, but I don't think this really makes much sense. The test as it stands doesn't check that anything "convolves correctly": it just checks that the optimised code behaves the same as the C reference code. Is there something that you think should be added? And, in that case, can you explain what class of bugs it would catch?

I'm afraid I don't understand your last comment at all. In particular, the comment about coverage seems a little odd: this patch does indeed run the C convolve code in a unit test, but doesn't add any meaningful test coverage for the C code (because the test assumes that the C code is right!)

If you think there should be a unit test checking that the C code performs convolution correctly, that's perfectly reasonable, but it is unrelated to this patch.

Rupert

Patch-set: 7
Reviewer: Gerrit User 7560 <7560@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>